requirements.txt
torch
torchvision
numpy
pandas
pillow
matplotlib
scikit-learn
pyyaml
tqdm
